Fixed eslint and moved clone to new file, made LarransChest force big or small and lots of other small changes
