+++ src/simulation/openables/BronzeHAMChest.ts
@@ -1,11 +1,7 @@
 import LootTable from '../../structures/LootTable';
 import SimpleOpenable from '../../structures/SimpleOpenable';
-import { rand } from '../../util';
-
-const rolls = rand(1, 5);
 
 /* Gem loots */
-
 // TODO: check back for wiki drop table update from another one in the repo
 const GemTable = new LootTable({ limit: 100 })
 	.add('Uncut sapphire', 1, 25)
@@ -17,7 +13,13 @@ const GemTable = new LootTable({ limit: 100 })
 	.add('Uncut diamond', 1, 5)
 	.add('Diamond', 1, 5);
 
-const BronzeHAMChestTable = new LootTable().every('Coins', [1, 400]).every(GemTable, rolls);
+const BronzeHAMChestTable = new LootTable()
+	.every('Coins', [1, 400])
+	.add(GemTable, 1, 1)
+	.add(GemTable, 2, 1)
+	.add(GemTable, 3, 1)
+	.add(GemTable, 4, 1)
+	.add(GemTable, 5, 1);
 
 export default new SimpleOpenable({
 	id: 8867,
