+++ src/simulation/openables/SteelHAMChest.ts
@@ -1,8 +1,5 @@
 import LootTable from '../../structures/LootTable';
 import SimpleOpenable from '../../structures/SimpleOpenable';
-import { rand } from '../../util';
-
-const rolls = rand(1, 5);
 
 /* Gem loots */
 // TODO: check back for wiki drop table update from another one in the repo
@@ -16,7 +13,13 @@ const GemTable = new LootTable({ limit: 100 })
 	.add('Uncut diamond', 1, 10)
 	.add('Diamond', 1, 10);
 
-const SteelHAMChestTable = new LootTable().every('Coins', [400, 800]).every(GemTable, rolls);
+const SteelHAMChestTable = new LootTable()
+	.every('Coins', [400, 800])
+	.add(GemTable, 1, 1)
+	.add(GemTable, 2, 1)
+	.add(GemTable, 3, 1)
+	.add(GemTable, 4, 1)
+	.add(GemTable, 5, 1);
 
 export default new SimpleOpenable({
 	id: 8866,
