+++ src/util/util.ts
@@ -191,23 +191,3 @@ export function addArrayOfNumbers(arr: number[]): number {
 export function JSONClone<O>(object: O): O {
 	return JSON.parse(JSON.stringify(object));
 }
-
-/**
- * @function
- * @description Deep clone a class instance.
- * @param instance The class instance you want to clone.
- * @returns A new cloned instance.
- */
-/*eslint-disable */
-export function clone(instance: any) {
-	return Object.assign(
-		Object.create(
-			// Set the prototype of the new object to the prototype of the instance.
-			// Used to allow new object behave like class instance.
-			Object.getPrototypeOf(instance)
-		),
-		// Prevent shallow copies of nested structures like arrays, etc
-		JSON.parse(JSON.stringify(instance))
-	);
-}
-/*eslint-enable */
